Fixing test_time reading
Previously, we referenced the test times based on the first reported test time,
assuming that the end line of one Excel file would exactly match the first line
of the subsequent file. That is not the case, so we cannot reference the test
times based on the first reported value on each file, and have to base ourselves
solely on the appropriate start time provided.